pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "python-dotenv",
    "typer[all]",
]

[project.entry-points.console_scripts]
ai = "gpt.main:app"
